requirements.txt
certifi==2016.9.26
dj-database-url==0.4.1
Django==1.10.3
django-crispy-forms==1.6.1
django-el-pagination==3.0.1
django-haystack==2.5.1
django-pagedown==0.1.1
django-registration-redux==1.4
elasticsearch==1.9.0
gunicorn==19.6.0
Pillow==3.4.2
psycopg2==2.6.2
pyelasticsearch==1.4
simplejson==3.10.0
six==1.10.0
urllib3==1.19.1
virtualenv==15.0.3
whitenoise==3.2.2
Whoosh==2.7.4
